+++ src/devices/RGBDFromRosTopic/RGBDFromRosTopic.h
@@ -23,6 +23,7 @@
 #include <yarp/os/Property.h>
 #include <yarp/dev/IRGBDSensor.h>
 #include <yarp/dev/RGBDSensorParamParser.h>
+#include <RGBDRosConversionUtils.h>
 
 #include <yarp/os/Node.h>
 #include <yarp/os/Subscriber.h>
@@ -65,40 +66,6 @@
  */
 
 typedef yarp::sig::ImageOf<yarp::sig::PixelFloat> depthImage;
-
-class commonImageProcessor:
-    public yarp::os::Subscriber<yarp::rosmsg::sensor_msgs::Image>
-{
-    protected:
-    yarp::sig::FlexImage   m_lastRGBImage;
-    depthImage             m_lastDepthImage;
-
-    protected:
-    std::mutex             m_port_mutex;
-    mutable yarp::os::Subscriber   <yarp::rosmsg::sensor_msgs::CameraInfo> m_subscriber_camera_info;
-    std::string            m_cameradata_topic_name;
-    std::string            m_camerainfo_topic_name;
-    mutable yarp::rosmsg::sensor_msgs::CameraInfo m_lastCameraInfo;
-    yarp::os::Stamp        m_lastStamp;
-    bool                   m_contains_rgb_data;
-    bool                   m_contains_depth_data;
-
-    public:
-    commonImageProcessor (std::string data_topic_name, std::string camera_info_topic_name);
-    virtual ~commonImageProcessor();
-    using yarp::os::Subscriber<yarp::rosmsg::sensor_msgs::Image>::onRead;
-    virtual void onRead(yarp::rosmsg::sensor_msgs::Image& v) override;
-
-    public:
-    size_t getWidth() const;
-    size_t getHeight() const;
-    bool getFOV(double& horizontalFov, double& verticalFov) const;
-    bool getIntrinsicParam(yarp::os::Property& intrinsic) const;
-
-    public:
-    bool getLastRGBData(yarp::sig::FlexImage& data, yarp::os::Stamp& stmp);
-    bool getLastDepthData(yarp::sig::ImageOf<yarp::sig::PixelFloat>& data, yarp::os::Stamp& stmp);
-};
 
 //---------------------------------------------------------------------------------------
 class RGBDFromRosTopic :
