+++ src/devices/RGBDSensorWrapper/CMakeLists.txt
@@ -15,8 +15,10 @@ if(NOT SKIP_RGBDSensorWrapper)
   yarp_add_plugin(yarp_RGBDSensorWrapper)
 
   target_sources(yarp_RGBDSensorWrapper PRIVATE RGBDSensorWrapper.cpp
-                                                RGBDSensorWrapper.h
-                                                rosPixelCode.h)
+                                                RGBDSensorWrapper.h)
+  target_sources(yarp_RGBDSensorWrapper PRIVATE $<TARGET_OBJECTS:RGBDRosConversionUtils>)
+
+  target_include_directories(yarp_RGBDSensorWrapper PRIVATE $<TARGET_PROPERTY:RGBDRosConversionUtils,INTERFACE_INCLUDE_DIRECTORIES>)
 
   target_link_libraries(yarp_RGBDSensorWrapper PRIVATE YARP::YARP_os
                                                        YARP::YARP_sig
